Application layers with Spring Boot 2.3 demo
This commit enables out-of-the-box support for application layers in the
repackaged archive. When this mode is enabled, a small utility is
included that lets us see the layers and manage them.

For instance, this command lists the available layers:

java -Djarmode=layertools -jar scribe/target/scribe-*.jar list

Out-of-the-box, four layers are defined:

* dependencies
* spring-boot-loader
* snapshot-dependencies
* application