Cloud Native Buildpacks demo
This commit removes the custom Dockerfile and leverage Spring Boot's
Cloud Native Buildpack support. Rather than using docker directly to
create the image, we can invoke a build plugin that invokes a builder.

An image can be built with Maven as follows:

./mvnw spring-boot:build-image -pl scribe

Or with Gradle:

./gradlew scribe:bootBuildImage

We can see that the image has transparently applied our layers
configuration that we've defined previously. Besides that, a number of
layers have been contributed based on the nature of the application. In
particular, the buildpack has detected the jar is a Spring Boot app and
a number of Spring Boot specific tasks were performed to optimize the
environment in which this app runs.

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe:0.0.1-SNAPSHOT