Cloud Native Buildpacks customization demo
Each buildpack is controlled by a number of environment variables that
can be set in the plugin's configuration. Spring Boot also detects the
target JVM version from the build and set the relevant properly
automatically.

This commit switches the project to use Java 11.

An image can be built with Maven as follows:

./mvnw spring-boot:build-image -pl scribe

Or with Gradle:

./gradlew scribe:bootBuildImage

We can see that the resulting image is now using JDK 11 while it was
previously using Java 8.

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe:0.0.1-SNAPSHOT