PostConstruct vs. ApplicationRunner demo
This commit showcases how the liveness and readiness probes can be used
by an application orchestrator. We've added an arbitrary delay in the
ApplicationRunner that loads our dictionary.

If we had this code in a PostConstruct, this would prevent the HTTP
connector to start accepting connection and the orchestrator may decide
that the application is faulty and kill it.

With this arrangement however, the application is live before the
ApplicationRunner is invoked. No traffic is routed to it just yet but it
is known as healthy. Once the runners complete, the application is live
and can service requests.

You can see that in action starting the app with two "watch" command if
you have that installed:

watch -n1 'http :8080/actuator/health/liveness'

watch -n1 'http :8080/actuator/health/readiness'