Health vs. Readiness demo
If we shutdown our `markdown-converter`, we can see that we can still
submit the form as we have a circuit breaker for the remote call that
fallbacks to a degraded mode.

Looking at `/actuator/health` we can see the general health of our
service is `out-of-service`, while the `readiness` probe is still `up`.

This is the right call considering that our access to the markdown
converter is covered. Using this technique, we can avoid shutting down
applications unnecessary when one component is not available.